+++ src/ideabox/policy/content/project.py
@@ -10,6 +10,7 @@ from plone.app.z3cform.widget import RichTextFieldWidget
 from plone.autoform import directives as form
 from plone.dexterity.browser import view
 from plone.dexterity.content import Container
+from plone.formwidget.geolocation.field import GeolocationField
 from plone.indexer.decorator import indexer
 from plone.namedfile.field import NamedBlobImage
 from plone.supermodel import model
@@ -53,6 +54,13 @@ class IProject(model.Schema):
 
     form.mode(original_author="hidden")
     original_author = schema.TextLine(title=_("Original author"), required=False)
+
+    geolocation = GeolocationField(
+        title=_('label_geolocation', default=u'Geolocation'),
+        description=_('help_geolocation',
+                      default=u'Click on the map to select a location, or '
+                              u'use the text input to search by address.'),
+        required=False)
 
 
 @implementer(IProject)