+++ src/ideabox/policy/form/project_submission.py
@@ -6,7 +6,6 @@ from ideabox.policy import logger
 from ideabox.policy.content.project import IProject
 from ideabox.policy.utils import execute_under_admin
 from plone import api
-from plone.namedfile.field import NamedBlobImage
 from z3c.form import button
 from z3c.form.field import Fields
 from z3c.form.form import Form
@@ -30,6 +29,7 @@ class ProjectSubmissionForm(Form):
         "project_district",
         "body",
         "project_image",
+        "geolocation",
         "original_author",
     )
     fields["project_theme"].widgetFactory = MultiSelect2FieldWidget