+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 3.4.8 (unreleased)
 ------------------
 
+- Add geolocation field to project content type
+  [boulch]
+
 - Move project_image (leadimage) to project content type
   [boulch]